requirements.txt
ImageHash
Pillow
PixivPy
SQLAlchemy
aiofiles
aiogram
aiohttp
aiohttp-socks
beautifulsoup4
emoji
lxml
mysqlclient
python-dateutil
pytumblr
vk-requests
